+++ cmn-data-svc-client/src/main/java/org/acumos/cds/client/CommonDataServiceRestClientMockImpl.java
@@ -37,8 +37,6 @@ import org.acumos.cds.domain.MLPNotebook;
 import org.acumos.cds.domain.MLPNotification;
 import org.acumos.cds.domain.MLPPasswordChangeRequest;
 import org.acumos.cds.domain.MLPPeer;
-import org.acumos.cds.domain.MLPPeerGroup;
-import org.acumos.cds.domain.MLPPeerSolAccMap;
 import org.acumos.cds.domain.MLPPeerSubscription;
 import org.acumos.cds.domain.MLPPipeline;
 import org.acumos.cds.domain.MLPProject;
@@ -54,7 +52,6 @@ import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPSolutionDeployment;
 import org.acumos.cds.domain.MLPSolutionDownload;
 import org.acumos.cds.domain.MLPSolutionFavorite;
-import org.acumos.cds.domain.MLPSolutionGroup;
 import org.acumos.cds.domain.MLPSolutionRating;
 import org.acumos.cds.domain.MLPSolutionRevision;
 import org.acumos.cds.domain.MLPTag;
@@ -179,15 +176,6 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 	private List<MLPTaskStepResult> stepResults;
 	private RestPageResponse<MLPTaskStepResult> searchStepResults;
 	private MLPUserNotifPref usrNotifPrefById = null;
-	private RestPageResponse<MLPPeerGroup> peerGroups;
-	private MLPPeerGroup peerGroup;
-	private RestPageResponse<MLPSolutionGroup> solutionGroups;
-	private MLPSolutionGroup solutionGroup;
-	private RestPageResponse<MLPPeer> peersInGroup;
-	private RestPageResponse<MLPSolution> solutionsInGroup;
-	private RestPageResponse<MLPPeerSolAccMap> peerSolutionGroupMaps;
-	private long peerSolutionAccess;
-	private List<MLPPeer> peerAccessList;
 	private RestPageResponse<MLPSolution> solutionsByDate;
 	private MLPTaskStepResult stepResultById;
 	private List<MLPCodeNamePair> pairs;
@@ -204,7 +192,6 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 	private RestPageResponse<MLPPublishRequest> publishRequests;
 	private RestPageResponse<MLPPublishRequest> searchPublishRequests;
 	private MLPPublishRequest publishRequest;
-	private RestPageResponse<MLPSolution> restrictedSolutions;
 	private long userNotificationCount;
 	private byte[] solutionImage;
 	private MLPCatalog catalog;
@@ -1481,156 +1468,6 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 	@Override
 	public void deleteTaskStepResult(long stepResultId) {
 		// How to mock?
-	}
-
-	public void setPeerGroups(RestPageResponse<MLPPeerGroup> peerGroups) {
-		this.peerGroups = peerGroups;
-	}
-
-	@Override
-	public RestPageResponse<MLPPeerGroup> getPeerGroups(RestPageRequest pageRequest) {
-		return peerGroups;
-	}
-
-	public void setPeerGroup(MLPPeerGroup peerGroup) {
-		this.peerGroup = peerGroup;
-	}
-
-	@Override
-	public MLPPeerGroup createPeerGroup(MLPPeerGroup peerGroup) {
-		return this.peerGroup;
-	}
-
-	@Override
-	public void updatePeerGroup(MLPPeerGroup peerGroup) {
-		this.peerGroup = peerGroup;
-	}
-
-	@Override
-	public void deletePeerGroup(Long peerGroupId) {
-		// How to mock?
-	}
-
-	public void setSolutionGroups(RestPageResponse<MLPSolutionGroup> solutionGroups) {
-		this.solutionGroups = solutionGroups;
-	}
-
-	@Override
-	public RestPageResponse<MLPSolutionGroup> getSolutionGroups(RestPageRequest pageRequest) {
-		return this.solutionGroups;
-	}
-
-	public void setSolutionGroup(MLPSolutionGroup solutionGroup) {
-		this.solutionGroup = solutionGroup;
-	}
-
-	@Override
-	public MLPSolutionGroup createSolutionGroup(MLPSolutionGroup solutionGroup) {
-		return this.solutionGroup;
-	}
-
-	@Override
-	public void updateSolutionGroup(MLPSolutionGroup solutionGroup) {
-		this.solutionGroup = solutionGroup;
-	}
-
-	@Override
-	public void deleteSolutionGroup(Long solutionGroupId) {
-		// How to mock?
-	}
-
-	public void setPeersInGroup(RestPageResponse<MLPPeer> peersInGroup) {
-		this.peersInGroup = peersInGroup;
-	}
-
-	@Override
-	public RestPageResponse<MLPPeer> getPeersInGroup(Long peerGroupId, RestPageRequest pageRequest) {
-		return peersInGroup;
-	}
-
-	@Override
-	public void addPeerToGroup(String peerId, Long peerGroupId) {
-		// How to mock?
-	}
-
-	@Override
-	public void dropPeerFromGroup(String peerId, Long peerGroupId) {
-		// How to mock?
-	}
-
-	public void setSolutionsInGroup(RestPageResponse<MLPSolution> solutionsInGroup) {
-		this.solutionsInGroup = solutionsInGroup;
-	}
-
-	@Override
-	public RestPageResponse<MLPSolution> getSolutionsInGroup(Long solutionGroupId, RestPageRequest pageRequest) {
-		return solutionsInGroup;
-	}
-
-	@Override
-	public void addSolutionToGroup(String solutionId, Long solutionGroupId) {
-		// How to mock?
-	}
-
-	@Override
-	public void dropSolutionFromGroup(String solutionId, Long solutionGroupId) {
-		// How to mock?
-	}
-
-	public void setPeerSolutionGroupMaps(RestPageResponse<MLPPeerSolAccMap> peerSolutionGroupMaps) {
-		this.peerSolutionGroupMaps = peerSolutionGroupMaps;
-	}
-
-	@Override
-	public RestPageResponse<MLPPeerSolAccMap> getPeerSolutionGroupMaps(RestPageRequest pageRequest) {
-		return peerSolutionGroupMaps;
-	}
-
-	@Override
-	public void mapPeerSolutionGroups(Long peerGroupId, Long solutionGroupId) {
-		// How to mock?
-	}
-
-	@Override
-	public void unmapPeerSolutionGroups(Long peerGroupId, Long solutionGroupId) {
-		// How to mock?
-	}
-
-	@Override
-	public void mapPeerPeerGroups(Long principalGroupId, Long resourceGroupId) {
-		// How to mock?
-	}
-
-	@Override
-	public void unmapPeerPeerGroups(Long principalGroupId, Long resourceGroupId) {
-		// How to mock?
-	}
-
-	public void setPeerSolutionAccess(long count) {
-		this.peerSolutionAccess = count;
-	}
-
-	@Override
-	public long checkRestrictedAccessSolution(String peerId, String solutionId) {
-		return peerSolutionAccess;
-	}
-
-	public void setPeerAccess(List<MLPPeer> peers) {
-		this.peerAccessList = peers;
-	}
-
-	@Override
-	public List<MLPPeer> getPeerAccess(String peerId) {
-		return peerAccessList;
-	}
-
-	public void setRestrictedSolutions(RestPageResponse<MLPSolution> restrictedSolutions) {
-		this.restrictedSolutions = restrictedSolutions;
-	}
-
-	@Override
-	public RestPageResponse<MLPSolution> findRestrictedAccessSolutions(String peerId, RestPageRequest pageRequest) {
-		return restrictedSolutions;
 	}
 
 	@Override