+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/MockClientTest.java
@@ -35,8 +35,6 @@ import org.acumos.cds.domain.MLPNotebook;
 import org.acumos.cds.domain.MLPNotification;
 import org.acumos.cds.domain.MLPPasswordChangeRequest;
 import org.acumos.cds.domain.MLPPeer;
-import org.acumos.cds.domain.MLPPeerGroup;
-import org.acumos.cds.domain.MLPPeerSolAccMap;
 import org.acumos.cds.domain.MLPPeerSubscription;
 import org.acumos.cds.domain.MLPPipeline;
 import org.acumos.cds.domain.MLPProject;
@@ -52,7 +50,6 @@ import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPSolutionDeployment;
 import org.acumos.cds.domain.MLPSolutionDownload;
 import org.acumos.cds.domain.MLPSolutionFavorite;
-import org.acumos.cds.domain.MLPSolutionGroup;
 import org.acumos.cds.domain.MLPSolutionRating;
 import org.acumos.cds.domain.MLPSolutionRevision;
 import org.acumos.cds.domain.MLPTag;
@@ -479,55 +476,6 @@ public class MockClientTest {
 		client.updateTaskStepResult(stepResult);
 		client.deleteTaskStepResult(1L);
 
-		RestPageResponse<MLPPeerGroup> peerGroups = new RestPageResponse<>();
-		client.setPeerGroups(peerGroups);
-		Assert.assertTrue(peerGroups == client.getPeerGroups(new RestPageRequest()));
-		MLPPeerGroup peerGroup = new MLPPeerGroup();
-		client.setPeerGroup(peerGroup);
-		Assert.assertTrue(peerGroup == client.createPeerGroup(peerGroup));
-		client.updatePeerGroup(peerGroup);
-		client.deletePeerGroup(0L);
-
-		RestPageResponse<MLPSolutionGroup> solGroups = new RestPageResponse<>();
-		client.setSolutionGroups(solGroups);
-		Assert.assertTrue(solGroups == client.getSolutionGroups(new RestPageRequest()));
-		MLPSolutionGroup solGroup = new MLPSolutionGroup();
-		client.setSolutionGroup(solGroup);
-		Assert.assertTrue(solGroup == client.createSolutionGroup(solGroup));
-		client.updateSolutionGroup(solGroup);
-		client.deleteSolutionGroup(0L);
-
-		RestPageResponse<MLPPeer> peersInGroup = new RestPageResponse<>();
-		client.setPeersInGroup(peersInGroup);
-		Assert.assertTrue(peersInGroup == client.getPeersInGroup(0L, new RestPageRequest()));
-		client.addPeerToGroup("peerId", 0L);
-		client.dropPeerFromGroup("peerId", 0L);
-
-		RestPageResponse<MLPSolution> solsInGroup = new RestPageResponse<>();
-		client.setSolutionsInGroup(solsInGroup);
-		Assert.assertTrue(solsInGroup == client.getSolutionsInGroup(0L, new RestPageRequest()));
-		client.addSolutionToGroup("peerId", 0L);
-		client.dropSolutionFromGroup("peerId", 0L);
-
-		RestPageResponse<MLPPeerSolAccMap> peerSolutionGroupMaps = new RestPageResponse<>();
-		client.setPeerSolutionGroupMaps(peerSolutionGroupMaps);
-		Assert.assertTrue(peerSolutionGroupMaps == client.getPeerSolutionGroupMaps(new RestPageRequest()));
-		client.mapPeerSolutionGroups(0L, 1L);
-		client.unmapPeerSolutionGroups(0L, 1L);
-		client.mapPeerPeerGroups(0L, 1L);
-		client.unmapPeerPeerGroups(0L, 1L);
-
-		long peerSolutionAccess = 1;
-		client.setPeerSolutionAccess(peerSolutionAccess);
-		Assert.assertTrue(peerSolutionAccess == client.checkRestrictedAccessSolution("peerId", "solutionId"));
-
-		List<MLPPeer> peerAccessList = new ArrayList<>();
-		client.setPeerAccess(peerAccessList);
-		Assert.assertTrue(peerAccessList == client.getPeerAccess("peerId"));
-		RestPageResponse<MLPSolution> restrictedSolutions = new RestPageResponse<>();
-		client.setRestrictedSolutions(restrictedSolutions);
-		Assert.assertTrue(restrictedSolutions == client.findRestrictedAccessSolutions("peerId", new RestPageRequest()));
-
 		List<String> members = new ArrayList<>();
 		client.setCompositeSolutionMembers(members);
 		Assert.assertTrue(members == client.getCompositeSolutionMembers("id"));
@@ -575,9 +523,9 @@ public class MockClientTest {
 		Assert.assertEquals(catalog, client.createCatalog(catalog));
 		client.updateCatalog(catalog);
 		client.deleteCatalog(catalog.getCatalogId());
-		RestPageResponse<MLPSolution> solsInCat = new RestPageResponse<>();
-		client.setSolutionsInCatalogs(solsInCat);
-		Assert.assertEquals(solsInCat, client.getSolutionsInGroup(0L, new RestPageRequest()));
+		RestPageResponse<MLPSolution> solsInCats = new RestPageResponse<>();
+		client.setSolutionsInCatalogs(solsInCats);
+		Assert.assertEquals(solsInCats, client.getSolutionsInCatalogs(new String[] {"abc"}, new RestPageRequest()));
 		client.addSolutionToCatalog("solId", "catId");
 		client.dropSolutionFromCatalog("solId", "catId");
 		client.getCatalogSolutionCount("catId");